refactor: extract shared schema methods into SchemaMethods mixin (#103)
Extract duplicated schema-related methods from Model and Schema modules
into a shared SchemaMethods module to eliminate code duplication and
prevent behavioral drift.

Extracted methods:
- ref_template
- json_schema
- build_json_schema (private)
- resolve_schema_uri (private)
- resolve_schema_id (private)
