+++ lib/easy_talk/schema.rb
@@ -129,6 +129,8 @@ module EasyTalk
 
     # Class methods for schema-only models.
     module ClassMethods
+      include SchemaMethods
+
       # Returns the schema for the model.
       #
       # @return [Hash] The schema for the model.
@@ -138,20 +140,6 @@ module EasyTalk
                     else
                       {}
                     end
-      end
-
-      # Returns the reference template for the model.
-      #
-      # @return [String] The reference template for the model.
-      def ref_template
-        "#/$defs/#{name}"
-      end
-
-      # Returns the JSON schema for the model.
-      #
-      # @return [Hash] The JSON schema for the model.
-      def json_schema
-        @json_schema ||= build_json_schema
       end
 
       # Define the schema for the model using the provided block.
@@ -204,53 +192,6 @@ module EasyTalk
       # @return [Hash] The built schema.
       def build_schema(schema_definition)
         Builders::ObjectBuilder.new(schema_definition).build
-      end
-
-      # Builds the final JSON schema with optional $schema and $id keywords.
-      #
-      # @return [Hash] The JSON schema.
-      def build_json_schema
-        result = schema.as_json
-        schema_uri = resolve_schema_uri
-        id_uri = resolve_schema_id
-
-        prefix = {}
-        prefix['$schema'] = schema_uri if schema_uri
-        prefix['$id'] = id_uri if id_uri
-
-        return result if prefix.empty?
-
-        prefix.merge(result)
-      end
-
-      # Resolves the schema URI from per-model setting or global config.
-      #
-      # @return [String, nil] The schema URI.
-      def resolve_schema_uri
-        model_version = @schema_definition&.schema&.dig(:schema_version)
-
-        if model_version
-          return nil if model_version == :none
-
-          Configuration::SCHEMA_VERSIONS[model_version] || model_version.to_s
-        else
-          EasyTalk.configuration.schema_uri
-        end
-      end
-
-      # Resolves the schema ID from per-model setting or global config.
-      #
-      # @return [String, nil] The schema ID.
-      def resolve_schema_id
-        model_id = @schema_definition&.schema&.dig(:schema_id)
-
-        if model_id
-          return nil if model_id == :none
-
-          model_id.to_s
-        else
-          EasyTalk.configuration.schema_id
-        end
       end
     end
   end
